requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==47.0.0
et_xmlfile==2.0.0
invoke==3.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==9.1.0
openpyxl==3.1.5
paramiko==4.0.0
pycparser==3.0
Pygments==2.20.0
PyNaCl==1.6.2
pyserial==3.5
python-dotenv==1.2.2
PyYAML==6.0.3
rich==15.0.0
ruamel.yaml==0.19.1
scp==0.15.0
textfsm==2.1.0
